+++ btreeg_iter.go
@@ -1,0 +1,238 @@
+package btree
+
+// Iter represents an iterator
+type IterG[T any] struct {
+	tr      *BTreeG[T]
+	mut     bool
+	locked  bool
+	seeked  bool
+	atstart bool
+	atend   bool
+	stack   []iterStackItemG[T]
+	item    T
+}
+
+type iterStackItemG[T any] struct {
+	n *node[T]
+	i int
+}
+
+// Iter returns a read-only iterator.
+// The Release method must be called finished with iterator.
+func (tr *BTreeG[T]) Iter() IterG[T] {
+	return tr.iter(false)
+}
+
+func (tr *BTreeG[T]) IterMut() IterG[T] {
+	return tr.iter(true)
+}
+
+// Seek to item greater-or-equal-to key.
+// Returns false if there was no item found.
+func (iter *IterG[T]) Seek(key T) bool {
+	if iter.tr == nil {
+		return false
+	}
+	iter.seeked = true
+	iter.stack = iter.stack[:0]
+	if iter.tr.root == nil {
+		return false
+	}
+	n := iter.tr.isoLoad(&iter.tr.root, iter.mut)
+	for {
+		i, found := iter.tr.find(n, key, nil, 0)
+		iter.stack = append(iter.stack, iterStackItemG[T]{n, i})
+		if found {
+			iter.item = n.items[i]
+			return true
+		}
+		if n.leaf() {
+			iter.stack[len(iter.stack)-1].i--
+			return iter.Next()
+		}
+		n = iter.tr.isoLoad(&(*n.children)[i], iter.mut)
+	}
+}
+
+// First moves iterator to first item in tree.
+// Returns false if the tree is empty.
+func (iter *IterG[T]) First() bool {
+	if iter.tr == nil {
+		return false
+	}
+	iter.atend = false
+	iter.atstart = false
+	iter.seeked = true
+	iter.stack = iter.stack[:0]
+	if iter.tr.root == nil {
+		return false
+	}
+	n := iter.tr.isoLoad(&iter.tr.root, iter.mut)
+	for {
+		iter.stack = append(iter.stack, iterStackItemG[T]{n, 0})
+		if n.leaf() {
+			break
+		}
+		n = iter.tr.isoLoad(&(*n.children)[0], iter.mut)
+	}
+	s := &iter.stack[len(iter.stack)-1]
+	iter.item = s.n.items[s.i]
+	return true
+}
+
+// Last moves iterator to last item in tree.
+// Returns false if the tree is empty.
+func (iter *IterG[T]) Last() bool {
+	if iter.tr == nil {
+		return false
+	}
+	iter.seeked = true
+	iter.stack = iter.stack[:0]
+	if iter.tr.root == nil {
+		return false
+	}
+	n := iter.tr.isoLoad(&iter.tr.root, iter.mut)
+	for {
+		iter.stack = append(iter.stack, iterStackItemG[T]{n, len(n.items)})
+		if n.leaf() {
+			iter.stack[len(iter.stack)-1].i--
+			break
+		}
+		n = iter.tr.isoLoad(&(*n.children)[len(n.items)], iter.mut)
+	}
+	s := &iter.stack[len(iter.stack)-1]
+	iter.item = s.n.items[s.i]
+	return true
+}
+
+// Release the iterator.
+func (iter *IterG[T]) Release() {
+	if iter.tr == nil {
+		return
+	}
+	if iter.locked {
+		iter.tr.unlock(iter.mut)
+		iter.locked = false
+	}
+	iter.stack = nil
+	iter.tr = nil
+}
+
+// Next moves iterator to the next item in iterator.
+// Returns false if the tree is empty or the iterator is at the end of
+// the tree.
+func (iter *IterG[T]) Next() bool {
+	if iter.tr == nil {
+		return false
+	}
+	if !iter.seeked {
+		return iter.First()
+	}
+	if len(iter.stack) == 0 {
+		if iter.atstart {
+			return iter.First() && iter.Next()
+		}
+		return false
+	}
+	s := &iter.stack[len(iter.stack)-1]
+	s.i++
+	if s.n.leaf() {
+		if s.i == len(s.n.items) {
+			for {
+				iter.stack = iter.stack[:len(iter.stack)-1]
+				if len(iter.stack) == 0 {
+					iter.atend = true
+					return false
+				}
+				s = &iter.stack[len(iter.stack)-1]
+				if s.i < len(s.n.items) {
+					break
+				}
+			}
+		}
+	} else {
+		n := iter.tr.isoLoad(&(*s.n.children)[s.i], iter.mut)
+		for {
+			iter.stack = append(iter.stack, iterStackItemG[T]{n, 0})
+			if n.leaf() {
+				break
+			}
+			n = iter.tr.isoLoad(&(*n.children)[0], iter.mut)
+		}
+	}
+	s = &iter.stack[len(iter.stack)-1]
+	iter.item = s.n.items[s.i]
+	return true
+}
+
+// Prev moves iterator to the previous item in iterator.
+// Returns false if the tree is empty or the iterator is at the beginning of
+// the tree.
+func (iter *IterG[T]) Prev() bool {
+	if iter.tr == nil {
+		return false
+	}
+	if !iter.seeked {
+		return false
+	}
+	if len(iter.stack) == 0 {
+		if iter.atend {
+			return iter.Last() && iter.Prev()
+		}
+		return false
+	}
+	s := &iter.stack[len(iter.stack)-1]
+	if s.n.leaf() {
+		s.i--
+		if s.i == -1 {
+			for {
+				iter.stack = iter.stack[:len(iter.stack)-1]
+				if len(iter.stack) == 0 {
+					iter.atstart = true
+					return false
+				}
+				s = &iter.stack[len(iter.stack)-1]
+				s.i--
+				if s.i > -1 {
+					break
+				}
+			}
+		}
+	} else {
+		n := iter.tr.isoLoad(&(*s.n.children)[s.i], iter.mut)
+		for {
+			iter.stack = append(iter.stack, iterStackItemG[T]{n, len(n.items)})
+			if n.leaf() {
+				iter.stack[len(iter.stack)-1].i--
+				break
+			}
+			n = iter.tr.isoLoad(&(*n.children)[len(n.items)], iter.mut)
+		}
+	}
+	s = &iter.stack[len(iter.stack)-1]
+	iter.item = s.n.items[s.i]
+	return true
+}
+
+// Item returns the current iterator item.
+func (iter *IterG[T]) Item() T {
+	return iter.item
+}
+
+// Items returns all the items in order.
+func (tr *BTreeG[T]) Items() []T {
+	return tr.items(false)
+}
+
+func (tr *BTreeG[T]) ItemsMut() []T {
+	return tr.items(true)
+}
+
+// Clear will delete all items.
+func (tr *BTreeG[T]) Clear() {
+	if tr.lock(true) {
+		defer tr.unlock(true)
+	}
+	tr.root = nil
+	tr.count = 0
+}
